クソでかコミット

+++ server/src/controller/commitment.go
@@ -33,7 +33,7 @@ func (c *CommitmentController) GetCount(w http.ResponseWriter, r *http.Request, 
 		return helper.CreateHTTPError(http.StatusInternalServerError, "failed to fetch count.")
 	}
 
-	rtn := map[string]int{
+	rtn := map[string]int64{
 		"count": count,
 	}
 	return helper.JSON(w, http.StatusOK, rtn)
